OHT-53 Refactor the plugin module
